refactor: share form + debug across locale routes
Extract SsrDebug component and getPageData server function into
shared modules. Both / and /$locale/ routes are now thin wrappers.
